add math formatting
